feat: new profile system that replaces Login
This touches a lot of things because util.Config is no longer
particularly useful. All the active session info is in
util.ActiveProfile. Further, the profile data can contain a default
workspace which allows the user to use 'workspace' commands without
provided a workspace name on the cli.

This also brings in 'tcnksm/go-input' as a new dependency.