In `conch workspace`, when figuring out WorkspaceUUID, stop defaulting to the active profile's workspace if the user passed in a workspace.
Previously, `conch workspace BAD get` would work if the active profile
had a workspace default. Instead of BAD, the user would receive info
about the profile's workspace instead of receiving an error.

A new error (`Workspace %s does not exist or you do not have permission
to access it`) is thrown if the user passes in a bad workspace
identifier.