Make the default textual output of `conch device :id get` a lot more useful.
There's a bit of an oddity here with the JSON output. Getting a device
is the most peformed action in the shell, sayeth server logs. Most
folks, near as I can tell, are using the JSON output.

The extended output provided by this change is vastly different from the
original output. Enough that I'd expect downstream tools to break. So,
for now, if --json is provided, we present the original Device
structure. To get the json of the new structure, the user needs to
provide `--json device :id get --extended`. The textual output will
*always* be the extended data.

This is a 'wat' label, I know, but back compat is kinda wat worthy
sometimes.