Improve DisplayDevices while combining it with TabelizeMinimalDevices
DisplayDevices has always accepted a list of conch.Device structures and
assumed that upstream filled in the right data, particularly location
data. That was not universally true. All but one of the callers had API
Device responses, not DetailedDevice, so were missing location data
entirely. `--full` would give the user incomplete data in that case.

Now, DisplayDevices checks for the location data and fetches it if need
be. The practical upshot here is that callers can just hand in partial
Devices instead of calling FillInDevice first. Just getting the location
is twice as fast, in my highly-anectodal experience.

The old code also had a concept of MinimalDevice, which was similar to
an API Device response but not quite and had the datacenter and rack
name shoved in there. The new code gets rid of MinimalDevice as a global
concept and pulls it into some back compat code instead.

The user facing change is that the "AZ" and "Rack" elements are no
longer provided in this minimal structure. I removed those fields
because (a) they're never populated without the caller first using
FillInDevice (b) having the rack name is pointless because we can't
actually look up a rack by its name currently.

The --full output includes the location data. Given that these are
Device + Location, they are not fully formed devices. This might be a
wat eventually but I think it's still an improvement on the current
situation.