breaking: HardwareProduct.Specification is now a proper data structure instead of a string.
Closes #187

This is a play in two acts.

Act I is in conch/structs, where HardwareProduct has a custom
UnmarshalJSON that unmarshals the API's hardware_product.specification
string into an interface{}. This lets the shell use it properly,
including outputing the field as a data structure in json.

However, the API still wants it back as a string for write operations.
Act II is over in conch/hardware, where we build a special outgoing
structure and marshal our HardwareProduct.Specification into a string.