in `workspace :id devices --full`, use the new Device.RackID field to let us fill in the table much much faster.
Local tests lower the time on `workspace us-east-1 devices --full` from
1min 24s to 7s. `workspace GLOBAL devices --full` goes from "I killed
the process because I got bored waiting" to 25s. Again, my local
environment.

One major caveat here is that the process of making this command usable
for large datasets means that I've severely reduced the amount of data
we're pulling for each device. As such, the amount of populated fields
in --json has dropped substantially.